WS3: relocate tidy3d engine to internal module, invert the dependency
The supported Tidy3DSolver adapter imported and instantiated the
DEPRECATED fdtd_solver_tidy3d to build its ComponentModeler - a
supported->deprecated dependency (the wrong direction).

Relocate-and-invert (strangler fig, zero behavior change):
- solver_tidy3d.py -> solvers/_tidy3d_engine.py, class renamed
  fdtd_solver_tidy3d -> _TidyEngine (internal, supported).
- solver_tidy3d.py is now a thin re-export so the deprecated public
  name keeps working through 0.5.x.
- the adapter's build() uses _TidyEngine directly; nothing supported
  reaches into deprecated code anymore.
Validated: mocked adapter build (fake tidy3d) 7/7; real tidy3d adapter
build produces the correct modeler (2 sims x 2 ports); full suite green;
F10 lazy-web-import guard repointed at the engine module.
